feat: AlarmController 추가

+++ src/main/java/com/ajousw/spring/domain/alarm/AlarmService.java
@@ -8,8 +8,9 @@ import com.ajousw.spring.domain.timetable.repository.Subject;
 import com.ajousw.spring.domain.timetable.repository.SubjectRepository;
 import com.ajousw.spring.web.controller.dto.alarm.AlarmCreateDto;
 import com.ajousw.spring.web.controller.dto.alarm.AlarmDeleteDto;
-import com.ajousw.spring.web.controller.dto.alarm.AlarmModifyDto;
-import com.ajousw.spring.web.controller.dto.alarm.AlarmSubjectDto;
+import com.ajousw.spring.web.controller.dto.alarm.AlarmDto;
+import com.ajousw.spring.web.controller.dto.alarm.AlarmUpdateDto;
+import com.ajousw.spring.web.controller.dto.timetable.SubjectDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -26,25 +27,6 @@ public class AlarmService {
     private final AlarmRepository alarmRepository;
     private final SubjectRepository subjectRepository;
     private final MemberRepository memberRepository;
-
-    private static AlarmSubjectDto getAlarmSubjectDto(Alarm alarm) {
-        Subject subject = alarm.getSubject();
-
-        return new AlarmSubjectDto(
-                alarm.getId(),
-                alarm.getMember().getId(),
-                subject.getSubjectId(),
-                alarm.getDay(),
-                alarm.getHour(),
-                alarm.getMinute(),
-                alarm.getAlarmGap(),
-                alarm.getIsAlarmOn(),
-                subject.getSubjectRealId(),
-                subject.getCode(),
-                subject.getName(),
-                subject.getProfessor()
-        );
-    }
 
     public void save(AlarmCreateDto alarmCreateDto, String email) {
         Member member = memberRepository.findByEmail(email).orElseThrow(() ->
@@ -66,32 +48,69 @@ public class AlarmService {
         alarmRepository.save(newAlarm);
     }
 
-    public List<AlarmSubjectDto> getAllAlarmByEmail(String email) {
+    public List<AlarmDto> getAllAlarmByEmail(String email) {
         List<Alarm> alarm = alarmRepository.findByMemberEmailFetch(email);
 
-        List<AlarmSubjectDto> dtoList = alarm.stream().map(a -> getAlarmSubjectDto(a))
+        List<AlarmDto> dtoList = alarm.stream().map(a -> getAlarmSubjectDto(a))
                 .collect(Collectors.toList());
 
         return dtoList;
     }
 
-    public AlarmSubjectDto getAlarmById(Long id) {
+    public AlarmDto getAlarmById(Long id) {
         Alarm alarm = alarmRepository.findByIdFetch(id).orElseThrow(() ->
                 new IllegalArgumentException("존재하지 않는 알람입니다."));
 
         return getAlarmSubjectDto(alarm);
     }
 
-    public void modifyAlarm(AlarmModifyDto alarmModifyDto) {
+    public void updateAlarm(AlarmUpdateDto alarmModifyDto, String email) {
         Alarm alarm = alarmRepository.findByIdFetch(alarmModifyDto.getAlarmId()).orElseThrow(() ->
                 new IllegalArgumentException("존재하지 않는 알람입니다."));
+
+        checkOwner(email, alarm);
 
         alarm.setAlarmOn(alarmModifyDto.getIsAlarmOn());
         alarm.setAlarmGap(alarm.getAlarmGap());
     }
 
-    public void deleteAlarm(AlarmDeleteDto alarmDeleteDto) {
+    public void deleteAlarm(AlarmDeleteDto alarmDeleteDto, String email) {
+        Alarm alarm = alarmRepository.findByIdFetch(alarmDeleteDto.getAlarmId()).orElseThrow(() ->
+                new IllegalArgumentException("존재하지 않는 알람입니다."));
+
+        checkOwner(email, alarm);
+
         alarmRepository.deleteById(alarmDeleteDto.getAlarmId());
     }
 
+    private void checkOwner(String email, Alarm alarm) {
+        Member member = memberRepository.findByEmail(email).orElseThrow(() ->
+                new IllegalArgumentException("존재하지 않는 유저입니다."));
+
+        if (alarm.getMember().getId() != member.getId()) {
+            throw new IllegalArgumentException("알람은 본인만 삭제할 수 있습니다.");
+        }
+    }
+
+
+    private AlarmDto getAlarmSubjectDto(Alarm alarm) {
+        Subject subject = alarm.getSubject();
+
+        return new AlarmDto(
+                alarm.getId(),
+                alarm.getMember().getId(),
+                alarm.getDay(),
+                alarm.getHour(),
+                alarm.getMinute(),
+                alarm.getAlarmGap(),
+                alarm.getIsAlarmOn(),
+                new SubjectDto(subject.getSubjectId(),
+                        subject.getSubjectRealId(),
+                        subject.getCode(),
+                        subject.getName(),
+                        subject.getProfessor(),
+                        subject.getTimeTable().getTableId(),
+                        null)
+        );
+    }
 }
